requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==4.1.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
phonenumbers==8.13.26
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
python-telegram-bot==20.7
sniffio==1.3.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
